requirements.txt
spotipy>=2.26.0
rich>=14.3.3
click>=8.3.1
Pillow>=12.1.1
ascii-magic>=2.7.4
python-dotenv>=1.2.2
requests>=2.32.5